+++ front-end/src/views/monaco-page/MonacoPage.tsx
@@ -125,7 +125,7 @@ const MonacoPage = () => {
               </Box>
               <br />
               <Box>
-                <WebGLApp appsessionId={sessionId} />
+                <WebGLApp appsessionId={sessionId} onMountChange={handleMountChange} />
               </Box>
               <Box mt={2}>
                 <DialogContent className="testdialog">
